Avoid using sd_uid_get_state() which is broken in Jessie
Disable 0.113/sessionmonitor-systemd-prepare-for-D-Bus-user-bus-mo.patch
and 0.113/sessionmonitor-systemd-Use-sd_uid_get_state-to-check.patch.
Those two patches add support for `systemd --user` sessions and the
D-Bus user model. For that they rely on sd_uid_get_state() which seems
to be broken in Jessie when logind is running under sysvinit. Since
those features are not yet needed in Jessie, do not apply those patches.

Closes: #825956